refactor: User와 Result 객체를 InputView, OutputView로 변경
