+++ src/main/java/lotto/view/GameMessage.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.view;
 
 public enum GameMessage {
     BUY("구입금액을 입력해 주세요."),
